moved all train.py statistics functions to new folder
